+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 Changelog
 =========
+## Unreleased
+### Added
+- Organization Type field type can customize the drop down first option
+
 ## 1.0.0-rc.21 - 2018-06-25
 ### Fixed
 - User's organization list should appear even if no organizations are present.